Update to latest infrastructure tools.
I accidentally broke website updating by deleting a personal GitHub token,
so I had to give the lambda a new token.

Sadly, this repo hasn't gotten a lot of attention from me. This is good,
in the sense that the infrastructure it set up has been quietly
working for a year or two now! But dependencies don't stop moving
because your work is complete. They're like sharks! They've gotta keep
swimming!

As a result, I had to update my dependencies: this meant (finally) tackling
the Terraform 0.13 migration, upgrading to the latest OnePassword CLI (which
is admittedly much nicer!), and upgrading all of the providers in this repo.

So, a yak shake of a yak shave of a yak shake, found while on my way to shave
another yak. This is my hobby!